requirements.txt
keyboard==0.13.5
pyperclip==1.8.2
pyttsx3==2.90
deep-translator==1.11.4
pyinstaller
pystray 
Pillow 